requirements.txt
Flask==2.2.2
werkzeug==2.2.2
line-bot-sdk==3.5.0
# Pillow
googletrans==3.1.0a0
bs4
imgbbpy
zhdate
zhconv
# pandas
firebase_admin
requests